requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
black==23.10.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
Django==4.2.6
djangorestframework==3.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
psycopg==3.1.12
psycopg-binary==3.1.12
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
